+++ irs-api/src/main/java/org/eclipse/tractusx/irs/configuration/DependenciesHealthIndicator.java
@@ -4,7 +4,7 @@
  *       2022: ISTOS GmbH
  *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *       2022,2023: BOSCH AG
- * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -49,15 +49,19 @@ import org.springframework.web.client.RestTemplate;
 @Component
 @Slf4j
 @ConditionalOnEnabledHealthIndicator("dependencies")
-class DependenciesHealthIndicator implements HealthIndicator {
+public class DependenciesHealthIndicator implements HealthIndicator {
 
     private final DependenciesHealthConfiguration dependenciesHealthConfiguration;
     private final RestTemplate restTemplate;
 
-    /* package */ DependenciesHealthIndicator(@Qualifier(NO_ERROR_REST_TEMPLATE) final RestTemplate noErrorRestTemplate,
+    public DependenciesHealthIndicator(@Qualifier(NO_ERROR_REST_TEMPLATE) final RestTemplate noErrorRestTemplate,
             final DependenciesHealthConfiguration dependenciesHealthConfiguration) {
         this.dependenciesHealthConfiguration = dependenciesHealthConfiguration;
         this.restTemplate = noErrorRestTemplate;
+    }
+
+    public DependenciesHealthConfiguration getConfig() {
+        return dependenciesHealthConfiguration;
     }
 
     @Override