+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/EdcControlPlaneClient.java
@@ -4,7 +4,7 @@
  *       2022: ISTOS GmbH
  *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *       2022,2023: BOSCH AG
- * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -111,7 +111,7 @@ public class EdcControlPlaneClient {
             querySpec.limit(limit);
         }
         return CatalogRequest.Builder.newInstance()
-                                     .providerUrl(providerUrl)
+                                     .counterPartyAddress(providerUrl)
                                      .protocol(DATASPACE_PROTOCOL_HTTP)
                                      .querySpec(querySpec.build())
                                      .build();
@@ -121,7 +121,7 @@ public class EdcControlPlaneClient {
             final String value) {
         final QuerySpec querySpec = QuerySpec.Builder.newInstance().filter(new Criterion(key, "=", value)).build();
         final var catalogRequest = CatalogRequest.Builder.newInstance()
-                                                         .providerUrl(providerConnectorUrl)
+                                                         .counterPartyAddress(providerConnectorUrl)
                                                          .protocol(DATASPACE_PROTOCOL_HTTP)
                                                          .querySpec(querySpec)
                                                          .build();