+++ irs-registry-client/src/test/java/org/eclipse/tractusx/irs/registryclient/discovery/DiscoveryFinderClientImplTest.java
@@ -4,7 +4,7 @@
  *       2022: ISTOS GmbH
  *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *       2022,2023: BOSCH AG
- * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -24,10 +24,13 @@
 package org.eclipse.tractusx.irs.registryclient.discovery;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.mockito.Mockito.when;
 
 import java.util.List;
 
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -36,46 +39,90 @@ import org.springframework.web.client.RestTemplate;
 
 @ExtendWith(MockitoExtension.class)
 class DiscoveryFinderClientImplTest {
+
     @Mock
     private RestTemplate restTemplate;
 
-    @Test
-    void shouldReturnEmptyListWhenDiscoveryReturnsNull() {
-        // Arrange
-        final String discoveryFinderUrl = "https://discovery.mock/api/administration/connectors/discovery/search";
-        final String providerUrl = "https://discovery.mock/api/administration/connectors/discovery";
-        final List<String> bpns = List.of("BPN123");
-        final DiscoveryFinderClientImpl discoveryFinderClient = new DiscoveryFinderClientImpl(discoveryFinderUrl,
-                restTemplate);
-        when(restTemplate.postForObject(providerUrl, bpns, EdcDiscoveryResult[].class)).thenReturn(null);
+    @Nested
+    class FindDiscoveryEndpointsTests {
 
-        // Act
-        final List<EdcDiscoveryResult> connectorEndpoints = discoveryFinderClient.findConnectorEndpoints(providerUrl,
-                bpns);
+        public static final String DISCOVERY_FINDER_URL = "dummyUrl";
 
-        // Assert
-        assertThat(connectorEndpoints).isNotNull().isEmpty();
+        private DiscoveryFinderClientImpl discoveryFinderClient;
+
+        @BeforeEach
+        void beforeEach() {
+            this.discoveryFinderClient = new DiscoveryFinderClientImpl(DISCOVERY_FINDER_URL, restTemplate);
+        }
+
+        @Test
+        void findDiscoveryEndpoints_shouldReturnResponseFromRestRequest() {
+            // Arrange
+            final DiscoveryResponse mockResponse = new DiscoveryResponse(
+                    List.of(new DiscoveryEndpoint("test-endpoint", "desc", "test-endpoint-addr", "docs", "resId")));
+
+            final var request = new DiscoveryFinderRequest(List.of("bpn"));
+            when(restTemplate.postForObject(DISCOVERY_FINDER_URL, request, DiscoveryResponse.class)).thenReturn(
+                    mockResponse);
+
+            // Act
+            final var response = discoveryFinderClient.findDiscoveryEndpoints(request);
+
+            // Assert
+            assertThat(response).isNotNull();
+            assertThat(response.endpoints().stream().map(DiscoveryEndpoint::endpointAddress)).containsExactly(
+                    "test-endpoint-addr");
+        }
+
+        @Test
+        void evictDiscoveryEndpointsCacheValues_shouldNotThrow() {
+            assertThatNoException().isThrownBy(discoveryFinderClient::evictDiscoveryEndpointsCacheValues);
+        }
+
     }
 
-    @Test
-    void shouldReturnEmptyListWhenDiscoveryReturnsEmpty() {
-        // Arrange
-        final String discoveryFinderUrl = "https://discovery.mock/api/administration/connectors/discovery/search";
-        final String providerUrl = "https://discovery.mock/api/administration/connectors/discovery";
-        final List<String> bpns = List.of("BPN123");
-        final DiscoveryFinderClientImpl discoveryFinderClient = new DiscoveryFinderClientImpl(discoveryFinderUrl,
-                restTemplate);
-        final EdcDiscoveryResult[] discoveryResponse = { new EdcDiscoveryResult("BPN123",
-                List.of("https://provider.edc"))
-        };
+    @Nested
+    class FindConnectorEndpointsTests {
 
-        when(restTemplate.postForObject(providerUrl, bpns, EdcDiscoveryResult[].class)).thenReturn(discoveryResponse);
+        @Test
+        void shouldReturnEmptyListWhenDiscoveryReturnsNull() {
+            // Arrange
+            final String discoveryFinderUrl = "https://discovery.mock/api/administration/connectors/discovery/search";
+            final String providerUrl = "https://discovery.mock/api/administration/connectors/discovery";
+            final List<String> bpns = List.of("BPN123");
+            final DiscoveryFinderClientImpl discoveryFinderClient = new DiscoveryFinderClientImpl(discoveryFinderUrl,
+                    restTemplate);
+            when(restTemplate.postForObject(providerUrl, bpns, EdcDiscoveryResult[].class)).thenReturn(null);
 
-        // Act
-        final List<EdcDiscoveryResult> connectorEndpoints = discoveryFinderClient.findConnectorEndpoints(providerUrl,
-                bpns);
+            // Act
+            final List<EdcDiscoveryResult> connectorEndpoints = discoveryFinderClient.findConnectorEndpoints(
+                    providerUrl, bpns);
 
-        // Assert
-        assertThat(connectorEndpoints).isNotNull().hasSize(1);
+            // Assert
+            assertThat(connectorEndpoints).isNotNull().isEmpty();
+        }
+
+        @Test
+        void shouldReturnEmptyListWhenDiscoveryReturnsEmpty() {
+            // Arrange
+            final String discoveryFinderUrl = "https://discovery.mock/api/administration/connectors/discovery/search";
+            final String providerUrl = "https://discovery.mock/api/administration/connectors/discovery";
+            final List<String> bpns = List.of("BPN123");
+            final DiscoveryFinderClientImpl discoveryFinderClient = new DiscoveryFinderClientImpl(discoveryFinderUrl,
+                    restTemplate);
+            final EdcDiscoveryResult[] discoveryResponse = { new EdcDiscoveryResult("BPN123",
+                    List.of("https://provider.edc"))
+            };
+
+            when(restTemplate.postForObject(providerUrl, bpns, EdcDiscoveryResult[].class)).thenReturn(
+                    discoveryResponse);
+
+            // Act
+            final List<EdcDiscoveryResult> connectorEndpoints = discoveryFinderClient.findConnectorEndpoints(
+                    providerUrl, bpns);
+
+            // Assert
+            assertThat(connectorEndpoints).isNotNull().hasSize(1);
+        }
     }
 }